Folded in changes from trunk (through and including revision 752)

+++ indri/branches/metzler/include/indri/QuerySpec.hpp
@@ -1754,7 +1754,7 @@ namespace indri {
         _priorName( priorName )
       {
       } 
-
+      
       std::string queryText() const {
         std::stringstream qtext;
         qtext << "#prior(" << _priorName << ")";
@@ -1805,7 +1805,7 @@ namespace indri {
         _children = unpacker.getScoredExtentVector( "children" );
       }
 
-      unsigned int _hashCode() const {
+      UINT64 _hashCode() const {
         UINT64 accumulator = 0;
 
         for( int i=0; i<_children.size(); i++ ) {
